Move document date logic to a mixin
